igb: Update igb to use a path similar to ixgbe to determine when to stop Tx
After reviewing the igb and ixgbe code I realized there are a few issues in
how the code is structured.  Specifically we are not checking the size of the
buffers being used in transmits and we are not using the same value to
determine when to stop or start a Tx queue.  As such the code is prone to be
buggy.

This patch makes it so that we have one value DESC_NEEDED that we will use for
starting and stopping the queue.  In addition we will check the size of
buffers being used when setting up a transmit so as to avoid a possible buffer
overrun if we were to receive a frame with a block of data larger than 32K in
skb->data.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
Tested-by: Aaron Brown <aaron.f.brown@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
[bwh: Backported to wheezy: Assume maximum size of fragment is PAGE_SIZE,
 not NETDEV_FRAG_PAGE_MAX_SIZE (which is not defined).  This is somewhat
 questionable as a forwarded skb could theoretically include a compound
 page fragment larger than IGB_MAX_DATA_PER_TXD, but the same problem
 exists upstream.]

Gbp-Pq: Topic features/all/igb
Gbp-Pq: Name 0108-igb-Update-igb-to-use-a-path-similar-to-ixgbe-to-det.patch