mm: Try harder to allocate vmemmap blocks
Hot-adding memory on x86_64 normally requires huge page allocation.
When this is done to a VM guest, it's usually because the system is
already tight on memory, so the request tends to fail.  Try to avoid
this by adding __GFP_REPEAT to the allocation flags.

Reported-and-tested-by: Bernhard Schmidt <Bernhard.Schmidt@lrz.de>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name mm-Try-harder-to-allocate-vmemmap-blocks.patch