x86/efi: Autoload efivars
efivars is generally useful to have on EFI systems, and in some cases
it may be impossible to load it after a kernel upgrade in order to
complete a boot loader update.  At the same time we don't want to
waste memory on non-EFI systems by making it built-in.

Instead, give it a module alias as if it's a platform driver, and
register a corresponding platform device whenever EFI runtime services
are available.  This should trigger udev to load it.


Gbp-Pq: Topic debian
Gbp-Pq: Name efi-autoload-efivars.patch