x86, mm: Avoid ABI change in 3.2.44
Commit 511ba86e1d38 'x86, mm: Patch out arch_flush_lazy_mmu_mode()
when running on bare metal', backported into 3.2.44, added a new
member to struct pv_mmu_ops.  The working instance of struct
pv_mmu_ops is exported to modules.  The structure is also embedded in
struct paravirt_patch_template, and offsets within this structure are
used to identify operations to be patched in at runtime
(pv-alternatives).

We therefore need to place the new function pointer at the end of
struct paravirt_patch_template and outside of struct pv_mmu_ops
altogether.

The new function pointer will be called by arch_flush_lazy_mmu_mode(),
which did not use to be exported and therefore should not be used by
modules.  Hopefully no-one wants to implement this OOT because they'll
be out of luck.


Gbp-Pq: Topic debian
Gbp-Pq: Name x86-mm-avoid-abi-change-in-3.2.44.patch