xfrm: Avoid ABI change in 3.2.31
Commit 3b59df46a449ec9975146d71318c4777ad086744 ('xfrm: Workaround
incompatibility of ESN and async crypto') added a new operation
'recheck' to struct xfrm_replay.  There is only one instance which
needs this to be different than 'check'.  So instead of adding the
operation, check whether we're using that instance and call the
recheck implementation directly.


Gbp-Pq: Topic debian
Gbp-Pq: Name xfrm-avoid-ABI-change-in-3.2.31.patch