inetpeer: Avoid ABI change in 3.2.52
inet_peer::gc_list and inet_peer::gc_rcu were added in the middle of
struct inet_peer.  This structure is only really used by the IPv4 and
IPv6 stacks, and the new fields only by the built-in inetpeer code,
but this still changed symbol versions for most of networking.

We do need to ensure that where IPv6 is modular we either maintain
binary compatibility or change the relevant symbol versions.  So move
the new fields to the end of the structure and hide them from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name inetpeer-avoid-abi-change-in-3.2.52.patch