phy/marvell: disable 4-port phys
See 723177 for more details. Briefly:

The Marvell PHY was originally disabled because it can cause networking
failures on some systems. According to Lennert Buytenhek this is because some
of the variants added did not share the same register layout. Since the known
cases are all 4-ports disable those variants (indicated by a 4 in the
penultimate position of the model name) until they can be audited for
correctness.


Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name disable-some-marvell-phys.patch