of: Create of_console_check() for selecting a console specified in /chosen
The devicetree has a binding for specifying the console device in the
/chosen node, but the kernel doesn't use it consistently. This change
adds an API for testing if a device node is a console, and adds a
preferred console entry if it is.

At the same time this patch removes the of_device_is_stdout_path() API
since it is unused.

Signed-off-by: Grant Likely <grant.likely@linaro.org>
Tested-by: Sascha Hauer <s.hauer@pengutronix.de>

Gbp-Pq: Topic features/all
Gbp-Pq: Name of-Create-of_console_check-for-selecting-a-console-s.patch