staging: rtl8723au: rtw_update_ht_cap23a(): Use correct mask when determining 40MHz channel location
commit 4dc5f8bab2663b902b9cb0be29198dbbf38aa3fb
Author: Jes Sorensen <Jes.Sorensen@redhat.com>
Date:   Tue May 27 13:43:12 2014 +0200

    staging: rtl8723au: Eliminate struct HT_info_element

    Yet another duplicate version of struct ieee80211_ht_operation

    Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>

introduced a bad error, applying the wrong value as the mask when
determining the location of the 40MHz channel, effectively making it
always determining it as not present. This caused hangs and delays
when running in 40MHz channel mode.

This should get it right.

Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

Gbp-Pq: Topic features/all/r8723au
Gbp-Pq: Name 0021-staging-rtl8723au-rtw_update_ht_cap23a-Use-correct-m.patch