dmi: Avoid ABI change in 3.2.59
struct dmi_strmatch was changed to reassign an unused bit of 'slot'
as 'exact_match'.  This is binary backward-compatible in the sense
that old modules continue to work with the new kernel, but new
modules that set this flag will cause an array overrun in an old
kernel because 'slot' is not bounds-checked.

Hide the structure change from genksyms, but rename the exported
functions using this structure so that new modules will depend on
them and cannot be loaded against an old kernel.


Gbp-Pq: Topic debian
Gbp-Pq: Name dmi-avoid-abi-change-in-3.2.59.patch