ALSA: Avoid ABI change for CVE-2014-4652 fix
Move the new member snd_card::user_ctl_lock to the end of the structure
and hide it from genksyms.  struct snd_card is always allocated by
snd_card_create() in sound/core/init.c and user_ctl_lock is used in
sound/core/control.c, and these are part of the same module so they
are sure to be consistent.


Gbp-Pq: Topic debian
Gbp-Pq: Name alsa-avoid-abi-change-for-cve-2014-4652-fix.patch