irq: Avoid ABI change in 3.2.61
Move the new fields to the end of struct irq_desc and hide them from
genksyms.  Although struct irq_desc is normally allocated as part of a
static array, access is abstracted through the irq_to_desc() function
so modules don't depend on the size of the structure.


Gbp-Pq: Topic debian
Gbp-Pq: Name irq-avoid-abi-change-in-3.2.61.patch