KVM: Allow KVM_CHECK_EXTENSION on the vm fd
Origin: https://git.kernel.org/linus/92b591a4c46b103ebd3fc0d03a084e1efd331253

The KVM_CHECK_EXTENSION is only available on the kvm fd today. Unfortunately
on PPC some of the capabilities change depending on the way a VM was created.

So instead we need a way to expose capabilities as VM ioctl, so that we can
see which VM type we're using (HV or PR). To enable this, add the
KVM_CHECK_EXTENSION ioctl to our vm ioctl portfolio.

Signed-off-by: Alexander Graf <agraf@suse.de>
Acked-by: Paolo Bonzini <pbonzini@redhat.com>

Gbp-Pq: Topic features/powerpc
Gbp-Pq: Name KVM-Allow-KVM_CHECK_EXTENSION-on-the-vm-fd.patch